requirements.txt
asgiref==3.10.0
boto3==1.40.70
botocore==1.40.70
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==4.2.7
django-cors-headers==4.3.1
django-ses==4.4.0
djangorestframework==3.14.0
fonttools==4.60.1
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
httplib2==0.31.0
idna==3.11
jmespath==1.0.1
kiwisolver==1.4.9
matplotlib==3.8.2
numpy==1.26.4
oauthlib==3.3.1
packaging==25.0
pandas==2.1.3
Pillow==10.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
reportlab==4.0.7
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
